Introduced several 5-line .cs files
