+++ src/Avalonia.OpenGL/GlDisplayType.cs
@@ -1,0 +1,8 @@
+namespace Avalonia.OpenGL
+{
+    public enum GlDisplayType
+    {
+        OpenGL2,
+        OpenGLES2
+    }
+}
