+++ src/Avalonia.OpenGL/IGlContext.cs
@@ -1,0 +1,8 @@
+namespace Avalonia.OpenGL
+{
+    public interface IGlContext
+    {
+        IGlDisplay Display { get; }
+        void MakeCurrent(IGlSurface surface);
+    }
+}
