+++ src/Avalonia.OpenGL/IGlDisplay.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Avalonia.OpenGL
 {
     public interface IGlDisplay
@@ -9,23 +7,5 @@ namespace Avalonia.OpenGL
         void ClearContext();
         int SampleCount { get; }
         int StencilSize { get; }
-    }
-    
-    public enum GlDisplayType
-    {
-        OpenGL2,
-        OpenGLES2
-    }
-
-    public interface IGlContext
-    {
-        IGlDisplay Display { get; }
-        void MakeCurrent(IGlSurface surface);
-    }
-
-    public interface IGlSurface : IDisposable
-    {
-        IGlDisplay Display { get; }
-        void SwapBuffers();
     }
 }
