+++ src/Avalonia.OpenGL/IGlPlatformSurfaceRenderTarget.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Avalonia.OpenGL
+{
+    public interface IGlPlatformSurfaceRenderTarget : IDisposable
+    {
+        IGlPlatformSurfaceRenderingSession BeginDraw();
+    }
+}
