+++ src/Avalonia.OpenGL/IGlSurface.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Avalonia.OpenGL
+{
+    public interface IGlSurface : IDisposable
+    {
+        IGlDisplay Display { get; }
+        void SwapBuffers();
+    }
+}
